+++ ftlib/modules/classifier.py
@@ -2,7 +2,7 @@ from typing import Tuple, Optional, List, Dict
 import torch.nn as nn
 import torch
 
-__all__ = ['Classifier', 'Classifier_with_heads']
+__all__ = ['Classifier']
 
 
 class Classifier(nn.Module):
@@ -74,83 +74,14 @@ class Classifier(nn.Module):
         predictions = self.head(f)
         return predictions, f
 
-    def get_parameters(self) -> List[Dict]:
+    def get_parameters(self, base_lr=1.0) -> List[Dict]:
         """A parameter list which decides optimization hyper-parameters,
             such as the relative learning rate of each layer
         """
         params = [
-            {"params": self.backbone.parameters(), "lr": 0.1 if self.finetune else 1.},
-            {"params": self.bottleneck.parameters(), "lr": 1.},
-            {"params": self.head.parameters(), "lr": 1.},
+            {"params": self.backbone.parameters(), "lr": 0.1 * base_lr if self.finetune else 1.0 * base_lr},
+            {"params": self.bottleneck.parameters(), "lr": 1.0 * base_lr},
+            {"params": self.head.parameters(), "lr": 1.0 * base_lr},
         ]
 
-        return params
-
-class Classifier_with_heads(nn.Module):
-    """A generic Classifier class for domain adaptation.
-
-    Args:
-        backbone (torch.nn.Module): Any backbone to extract 2-d features from data
-        num_classes (int): Number of classes
-        bottleneck (torch.nn.Module, optional): Any bottleneck layer. Use no bottleneck by default
-        bottleneck_dim (int, optional): Feature dimension of the bottleneck layer. Default: -1
-        head (torch.nn.Module, optional): Any classifier head. Use :class:`torch.nn.Linear` by default
-        finetune (bool): Whether finetune the classifier or train from scratch. Default: True
-
-    .. note::
-        Different classifiers are used in different domain adaptation algorithms to achieve better accuracy
-        respectively, and we provide a suggested `Classifier` for different algorithms.
-        Remember they are not the core of algorithms. You can implement your own `Classifier` and combine it with
-        the domain adaptation algorithm in this algorithm library.
-
-    .. note::
-        The learning rate of this classifier is set 10 times to that of the feature extractor for better accuracy
-        by default. If you have other optimization strategies, please over-ride :meth:`~Classifier.get_parameters`.
-
-    Inputs:
-        - x (tensor): input data fed to `backbone`
-
-    Outputs:
-        - predictions: classifier's predictions
-        - features: features after `bottleneck` layer and before `head` layer
-
-    Shape:
-        - Inputs: (minibatch, *) where * means, any number of additional dimensions
-        - predictions: (minibatch, `num_classes`)
-        - features: (minibatch, `features_dim`)
-
-    """
-
-    def __init__(self, backbone: nn.Module, num_classes: int,  source_head: Optional[nn.Module] = None, finetune=True):
-        super(Classifier_with_heads, self).__init__()
-        self.backbone = backbone
-        self.num_classes = num_classes
-        if source_head is not None:
-            self.source_head = source_head
-
-        self._features_dim = self.backbone.out_features
-        self.head = nn.Linear(self._features_dim, num_classes)
-        self.finetune = finetune
-
-    @property
-    def features_dim(self) -> int:
-        """The dimension of features before the final `head` layer"""
-        return self._features_dim
-
-    def forward(self, x: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
-        """"""
-        f = self.backbone(x)
-        source_prediction = self.source_head(f)
-        prediction = self.head(f)
-        return source_prediction, prediction, f
-
-    def get_parameters(self) -> List[Dict]:
-        """A parameter list which decides optimization hyper-parameters,
-            such as the relative learning rate of each layer
-        """
-        params = [
-            {"params": self.backbone.parameters(), "lr": 0.1 if self.finetune else 1.},
-            {"params": self.source_head.parameters(), "lr": 0.1},
-            {"params": self.head.parameters(), "lr": 1.},
-        ]
         return params
